Move old `C++` examples to deprecated
