+++ examples/cpp/deprecated.cpp
@@ -9,7 +9,7 @@
 #include <vector>
 
 int main(int argc, char* argv[]) {
-    std::string filename = "drell-yan-rap-ll.pineappl.lz4";
+    std::string filename = "drell-yan-rap-ll-deprecated.pineappl.lz4";
     std::string pdfset = "NNPDF31_nlo_as_0118_luxqed";
 
     switch (argc) {
