+++ examples/cpp/display-channels-deprecated.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 
 int main(int argc, char* argv[]) {
-    std::string filename = "drell-yan-rap-ll.pineappl.lz4";
+    std::string filename = "drell-yan-rap-ll-deprecated.pineappl.lz4";
 
     switch (argc) {
     case 2:
@@ -23,24 +23,24 @@ int main(int argc, char* argv[]) {
     auto* grid = pineappl_grid_read(filename.c_str());
 
     // extract all channels
-    auto* channels = pineappl_grid_channels(grid);
+    auto* channels = pineappl_grid_lumi(grid);
 
     // how many channels are there?
-    auto channel_count = pineappl_channels_count(channels);
+    auto channel_count = pineappl_lumi_count(channels);
 
     for (std::size_t channel = 0; channel != channel_count; ++channel) {
         // print channel index
         std::cout << std::setw(4) << channel << ' ';
 
         // how many partonic combinations does this channel have?
-        auto combinations = pineappl_channels_combinations(channels, channel);
+        auto combinations = pineappl_lumi_combinations(channels, channel);
 
         std::vector<double> factors(combinations);
         std::vector<int> pids(2 * combinations);
 
         // read out the channel with index given by `channel`, writing the particle identifiers into
         // `pids` and the corresponding factors into `factors`
-        pineappl_channels_entry(channels, channel, pids.data(), factors.data());
+        pineappl_lumi_entry(channels, channel, pids.data(), factors.data());
 
         for (std::size_t combination = 0; combination != combinations; ++combination) {
             auto factor = factors.at(combination);
@@ -60,6 +60,6 @@ int main(int argc, char* argv[]) {
     }
 
     // release memory
-    pineappl_channels_delete(channels);
+    pineappl_lumi_delete(channels);
     pineappl_grid_delete(grid);
 }
